# services/file/api/openapi.yaml
openapi: 3.0.3
info:
  title: File Service Internal API
  version: 0.2.0
  description: Internal base file-object API used by owner services such as knowledge and document. Public frontend contracts remain in docs/services/gateway/api/openapi.yaml and must be reached only through gateway /api/v1/**. This contract intentionally avoids knowledge-base, report-template, report-material, and report-file business fields; those are owned by their corresponding services. The current code still keeps knowledge-document shaped MVP routes for compatibility, but new integrations should use /internal/v1/files/**.
servers:
  - url: http://localhost:8082
tags:
  - name: health
  - name: files
paths:
  /healthz:
    get:
      tags: [health]
      summary: Liveness check
      operationId: getHealth
      responses:
        '200':
          description: File service is alive.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /readyz:
    get:
      tags: [health]
      summary: Readiness check
      operationId: getReady
      responses:
        '200':
          description: File service is ready.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
        '503':
          description: File service is not ready because a runtime dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
  /internal/v1/files:
    post:
      tags: [files]
      summary: Create base file object
      operationId: createFile
      security:
        - ServiceTokenAuth: []
      description: Stores a raw file object and returns only base file metadata. Caller-owned business metadata such as knowledgeBaseId, reportId, templateId, materialId, reportFileId, tags, processing state, and ACLs must be stored by the caller's owner service.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateFileRequest'
      responses:
        '201':
          description: File object created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
  /internal/v1/files/{fileId}:
    get:
      tags: [files]
      summary: Get base file metadata
      operationId: getFile
      security:
        - ServiceTokenAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [files]
      summary: Delete base file object
      operationId: deleteFile
      security:
        - ServiceTokenAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '204':
          description: File deleted or marked deleted.
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /internal/v1/files/{fileId}/content:
    get:
      tags: [files]
      summary: Get original file content
      operationId: getFileContent
      security:
        - ServiceTokenAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Original file stream.
          headers:
            X-Request-Id:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    ServiceTokenAuth:
      type: apiKey
      in: header
      name: X-Service-Token
  parameters:
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Standard service error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    HealthResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [service, status]
          properties:
            service:
              type: string
              example: file
            status:
              type: string
              example: ready
        requestId:
          type: string
    ReadyResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [service, status, metadataBackend, storageBackend, serviceTokenConfigured, dependencies]
          properties:
            service:
              type: string
              example: file
            status:
              type: string
              enum: [ready, not_ready]
            metadataBackend:
              type: string
              enum: [memory, postgres]
            storageBackend:
              type: string
              example: local
            serviceTokenConfigured:
              type: boolean
            dependencies:
              type: array
              items:
                type: object
                required: [name, status]
                properties:
                  name:
                    type: string
                    example: postgres
                  status:
                    type: string
                    enum: [ready, unavailable, not_configured]
                  message:
                    type: string
        requestId:
          type: string
    CreateFileRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        checksumSha256:
          type: string
          description: Optional caller-computed SHA-256 checksum. File service may compute it when omitted.
      description: Multipart request for creating a base file object. Do not include business owner fields here.
    FileObject:
      type: object
      required: [id, filename, contentType, sizeBytes, createdAt]
      properties:
        id:
          type: string
          description: Internal file id for owner services. Frontend public responses for report templates, materials, and report files must not expose this internal id.
        filename:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        checksumSha256:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
    FileResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/FileObject'
        requestId:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum: [validation_error, unauthorized, forbidden, not_found, conflict, rate_limited, dependency_error, internal_error]
            message:
              type: string
            requestId:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string